+++ tests/repository/test_yum_repository.py
@@ -1,4 +1,6 @@
-from pubtools.pulplib import Repository, YumRepository
+import pytest
+
+from pubtools.pulplib import Repository, YumRepository, PulpException
 
 
 def test_from_data_gives_yum_repository():
@@ -76,3 +78,39 @@ def test_populate_attrs():
     assert repo.ubi_population
     assert repo.content_set == "fake_content_set"
     assert repo.ubi_config_version == "fake_ubi_config_version"
+
+
+def test_related_repositories(client, requests_mocker):
+    """test Repository.relate_*_repository returns expected objects"""
+
+    repo_binary_test = YumRepository(id="binary-repo", relative_url="some/repo/os")
+    repo_binary_test.__dict__["_client"] = client
+
+    requests_mocker.post(
+        "https://pulp.example.com/pulp/api/v2/repositories/search/",
+        [{"json": [{"id": "repo_debug"}]}, {"json": [{"id": "repo_source"}]}],
+    )
+
+    # Request for binary repo should return identical object
+    assert repo_binary_test is repo_binary_test.related_binary_repository
+    # Requests for debug and source repositories return correct repositories
+    assert repo_binary_test.related_debug_repository.id == "repo_debug"
+    assert repo_binary_test.related_source_repository.id == "repo_source"
+
+
+def test_related_repositories_not_found(client, requests_mocker):
+    """test Repository.relate_*_repository raises when repository is not found"""
+
+    repo_binary_test = YumRepository(id="binary-repo", relative_url="some/repo/os")
+    repo_binary_test.__dict__["_client"] = client
+
+    requests_mocker.post(
+        "https://pulp.example.com/pulp/api/v2/repositories/search/", json=[]
+    )
+
+    # It should raise
+    with pytest.raises(PulpException) as error:
+        repo_binary_test.related_debug_repository.result()
+
+    # It should explain the problem
+    assert "Repository relative_url=some/repo/debug was not found" in str(error.value)